chnage refund validatoin from integer to numeric

+++ app/Http/Requests/RefundRequest.php
@@ -33,7 +33,7 @@ class RefundRequest extends FormRequest
     {
         return [
             'type' => ['nullable'],
-            'amount' => ['nullable', 'required_if:type,partial_refund', new AmountPartialRefund($this->id), 'integer', 'gt:0', new AmountPartialRefundGTBalance($this->id)],
+            'amount' => ['nullable', 'required_if:type,partial_refund', new AmountPartialRefund($this->id), 'numeric', 'gt:0', new AmountPartialRefundGTBalance($this->id)],
         ];
     }
